// src/models/buttonType.ts
enum ButtonType {
  // Filled Black
  FILLED_BIG = "btn-filled-black btn-big",
  FILLED_MEDIUM = "btn-filled-black btn-medium",
  FILLED_SMALL = "btn-filled-black btn-small",
  // Filled White
  FILLED_WHITE_BIG = "btn-filled-white btn-big",
  FILLED_WHITE_MEDIUM = "btn-filled-white btn-medium",
  FILLED_WHITE_SMALL = "btn-filled-white btn-small",
  // Outlined
  OUTLINED_BIG = "btn-outlined-black btn-big",
  OUTLINED_MEDIUM = "btn-outlined-black btn-medium",
  OUTLINED_SMALL = "btn-outlined-black btn-small",
  // Outlined White
  OUTLINED_WHITE_BIG = "btn-outlined-white btn-big",
  OUTLINED_WHITE_MEDIUM = "btn-outlined-white btn-medium",
  OUTLINED_WHITE_SMALL = "btn-outlined-white btn-small",
  // Link White
  LINK_WHITE_BIG = "btn-link-white btn-big",
  LINK_WHITE_MEDIUM = "btn-link-white btn-medium",
  LINK_WHITE_SMALL = "btn-link-white btn-small",
  // Link Green
  LINK_GREEN_BIG = "btn-link-green btn-big",
  LINK_GREEN_MEDIUM = "btn-link-green btn-medium",
  LINK_GREEN_SMALL = "btn-link-green btn-small",
  // Link Black
  LINK_BLACK_BIG = "btn-link-black btn-big",
  LINK_BLACK_MEDIUM = "btn-link-black btn-medium",
  LINK_BLACK_SMALL = "btn-link-black btn-small",
  // Filled Black Icon Rounded
  FILLED_BIG_ROUNDED = "btn-filled-black btn-big-icon rounded",
  FILLED_MEDIUM_ROUNDED = "btn-filled-black btn-medium-icon rounded",
  FILLED_SMALL_ROUNDED = "btn-filled-black btn-small-icon rounded",
  // Filled White Icon Rounded
  FILLED_WHITE_BIG_ROUNDED = "btn-filled-white btn-big-icon rounded",
  FILLED_WHITE_MEDIUM_ROUNDED = "btn-filled-white btn-medium-icon rounded",
  FILLED_WHITE_SMALL_ROUNDED = "btn-filled-white btn-small-icon rounded",
  // Filled Aqua Icon Rounded
  FILLED_AQUA_BIG_ROUNDED = "btn-filled-aqua btn-big-icon rounded",
  FILLED_AQUA_MEDIUM_ROUNDED = "btn-filled-aqua btn-medium-icon rounded",
  FILLED_AQUA_SMALL_ROUNDED = "btn-filled-aqua btn-small-icon rounded",
  // Filled Aqua Icon Rounded
  OUTLINED_AQUA_BIG_ROUNDED = "btn-outlined-green btn-big-icon rounded",
  OUTLINED_AQUA_MEDIUM_ROUNDED = "btn-outlined-green btn-medium-icon rounded",
  OUTLINED_AQUA_SMALL_ROUNDED = "btn-outlined-green btn-small-icon rounded",
  // Filled Black Icon Square
  FILLED_BIG_SQUARE = "btn-filled-black btn-big-icon square",
  FILLED_MEDIUM_SQUARE = "btn-filled-black btn-medium-icon square",
  FILLED_SMALL_SQUARE = "btn-filled-black btn-small-icon square",
  // Filled White Icon Square
  FILLED_WHITE_BIG_SQUARE = "btn-filled-white btn-big-icon square",
  FILLED_WHITE_MEDIUM_SQUARE = "btn-filled-white btn-medium-icon square",
  FILLED_WHITE_SMALL_SQUARE = "btn-filled-white btn-small-icon square",
  // Filled Aqua Icon Square
  FILLED_AQUA_BIG_SQUARE = "btn-filled-aqua btn-big-icon square",
  FILLED_AQUA_MEDIUM_SQUARE = "btn-filled-aqua btn-medium-icon square",
  FILLED_AQUA_SMALL_SQUARE = "btn-filled-aqua btn-small-icon square",
  // Filled Aqua Icon Square
  OUTLINED_AQUA_BIG_SQUARE = "btn-outlined-green btn-big-icon square",
  OUTLINED_AQUA_MEDIUM_SQUARE = "btn-outlined-green btn-medium-icon square",
  OUTLINED_AQUA_SMALL_SQUARE = "btn-outlined-green btn-small-icon square",
}

export default ButtonType;